requirements/base.txt
sphinx==4.4.0
sphinx_rtd_theme==0.5.2
myst-parser==0.17.0
mdit-py-plugins
sphinx-copybutton==0.5.0
python-slugify==6.1.1

beautifulsoup4==4.11.1
markdownify==0.11.1
